CommonTextField の normalSupportingText 引数を削除
