+++ app/src/main/java/com/example/karaoke_note/ui/component/CommonTextField.kt
@@ -23,8 +23,10 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.*
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
@@ -37,6 +39,72 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.PopupProperties
 
+
+// スコア欄をチェックするための関数
+fun getNumberOfDecimalPoints(str: String): Int {
+    return str.count { it == '.' }
+}
+fun getIntegerPartOfScore(str: String): String {
+    val strArr = str.split(".").map { it.trim() }
+    return strArr[0]
+}
+fun getDecimalPartOfScore(str: String): String {
+    val strArr = str.split(".").map { it.trim() }
+    return strArr[1]
+}
+fun getSupportingTextForScoreField(
+    score: String,
+    isPlanning: Boolean,
+): String {
+    var message = ""
+
+    // スコア欄は予約フラグとの兼ね合いがある
+    if (!isPlanning) {
+        if (score.isBlank()) {
+            message = "Required."
+        }
+        else {
+            // スコア欄固有のチェック
+            // 小数が1つ以外のとき
+            if (getNumberOfDecimalPoints(score) != 1) {
+                message = "Invalid value."
+            }
+            else {
+                // 整数部、小数部をチェック
+                // 小数部は桁数が3桁に足りない場合は末尾に 0 を補充する
+                val strIntegerPart = getIntegerPartOfScore(score)
+                val integerPart: Int? = strIntegerPart.toIntOrNull()
+                val strDecimalPart: String = getDecimalPartOfScore(score)
+                val decimalPart: Int? = strDecimalPart.toIntOrNull()
+
+                // 何かしら値が入っているとき
+                if (integerPart != null && decimalPart != null) {
+                    // 100 点以上の扱い
+                    if ((integerPart > 100) || ((integerPart == 100) && (decimalPart != 0))) {
+                        message = "Invalid value."
+                    }
+                    else if ((integerPart < 0)) {
+                        // 負の数の扱い
+                        message = "Invalid value."
+                    }
+                    else {
+                        // 99.1234 などの変な形の時
+                        if (strDecimalPart.length != 3) {
+                            message = "Invalid value."
+                        }
+                    }
+                }
+                else {
+                    // .123 や ,.,,, みたいな形の時
+                    message = "Invalid value."
+                }
+            }
+        }
+    }
+
+    return message
+}
+
 @OptIn(ExperimentalFoundationApi::class)
 @ExperimentalMaterial3Api
 @Composable
@@ -48,7 +116,6 @@ fun CommonTextField(
     isEmptyAllowed: Boolean,      // TextField を空欄にすることを許可するかどうか
     singleLine: Boolean,
     fontSize: Int,
-    normalSupportingText: String,
     errorSupportingText: String,
     isEnabled: Boolean,
     keyboardType: KeyboardType,
@@ -67,7 +134,6 @@ fun CommonTextField(
         autoCompleteSuggestions.filter { it.startsWith(textFieldValue.text, ignoreCase = true) }
     }
     var isFocused by remember { mutableStateOf(false) }
-
 
     LaunchedEffect(initValue) {
         if (textFieldValue.text != initValue) {
@@ -104,12 +170,6 @@ fun CommonTextField(
                         text = errorSupportingText,
                         color = MaterialTheme.colorScheme.error
                     )
-                } else {
-                    Text(
-                        modifier = Modifier.fillMaxWidth(),
-                        text = normalSupportingText,
-                        color = MaterialTheme.colorScheme.background
-                    )
                 }
             },
             enabled = isEnabled,
@@ -143,6 +203,7 @@ fun CommonTextField(
                 }
             },
         )
+
         val alreadyComplete = filteredAutoCompleteSuggestions.size == 1 && filteredAutoCompleteSuggestions[0] == textFieldValue.text
         if (isFocused && filteredAutoCompleteSuggestions.isNotEmpty() && !alreadyComplete) {
             DropdownMenu(
